Create README.md
Model deployment is the process of putting machine learning models into production. The reason why we deploy machine learning models is to make them available for users, developers or systems to interact with the applications thereby making good business decision with it’s main goal of solving a problem
There are many ways to deploy machine Learning models, examples are StreamLit, AWS(Amazon Web Services), Flask to name a few. In this project, we will be using flask to deploy models on to the web.

The main idea of this project is to develop a machine learning algorithm using one of the common toy data set called the iris data and then deploying it on a webserver using Flask.